+++ features/overlays/rareCatchesTracker.js
@@ -3,21 +3,27 @@ import * as triggers from '../../constants/triggers';
 import * as seaCreatures from '../../constants/seaCreatures';
 import { persistentData } from "../../data/data";
 import { overlayCoordsData } from "../../data/overlayCoords";
-import { formatNumberWithSpaces, fromUppercaseToCapitalizedFirstLetters, isDoubleHook, pluralize } from '../../utils/common';
+import { formatNumberWithSpaces, fromUppercaseToCapitalizedFirstLetters, isDoubleHook, isInFishingWorld, pluralize } from '../../utils/common';
 import { WHITE, GOLD, BOLD, YELLOW, GRAY, RED } from "../../constants/formatting";
 import { RARE_CATCH_TRIGGERS } from "../../constants/triggers";
 import { getWorldName, hasFishingRodInHotbar, isInSkyblock } from "../../utils/playerState";
-import { KUUDRA } from "../../constants/areas";
 import { createButtonsDisplay, toggleButtonsDisplay } from "../../utils/overlays";
+import { registerIf } from "../../utils/registers";
 
 triggers.RARE_CATCH_TRIGGERS.forEach(entry => {
-    register("Chat", (event) => {
-        const isDoubleHooked = isDoubleHook();
-        trackCatch({ seaCreature: entry.seaCreature, rarityColorCode: entry.rarityColorCode, isDoubleHook: isDoubleHooked });
-    }).setCriteria(entry.trigger).setContains();
+    registerIf(
+        register("Chat", (event) => {
+            const isDoubleHooked = isDoubleHook();
+            trackCatch({ seaCreature: entry.seaCreature, rarityColorCode: entry.rarityColorCode, isDoubleHook: isDoubleHooked });
+        }).setCriteria(entry.trigger).setContains(),
+        () => settings.rareCatchesTrackerOverlay && isInSkyblock() && isInFishingWorld(getWorldName())
+    );
 });
 
-register('renderOverlay', () => renderRareCatchTrackerOverlay());
+registerIf(
+    register('renderOverlay', () => renderRareCatchTrackerOverlay()),
+    () => settings.rareCatchesTrackerOverlay && isInSkyblock() && isInFishingWorld(getWorldName())
+);
 
 register("gameUnload", () => {
     if (settings.rareCatchesTrackerOverlay && settings.resetRareCatchesTrackerOnGameClosed && persistentData.totalRareCatches > 0) {
@@ -67,7 +73,7 @@ export function resetRareCatchesTracker(isConfirmed) {
 
 function trackCatch(options) {
     try {
-        if (!settings.rareCatchesTrackerOverlay || !isInSkyblock()) {
+        if (!settings.rareCatchesTrackerOverlay || !isInSkyblock() || !isInFishingWorld(getWorldName())) {
             return;
         }
     
@@ -105,7 +111,7 @@ function renderRareCatchTrackerOverlay() {
     if (!settings.rareCatchesTrackerOverlay ||
         !Object.entries(persistentData.rareCatches).length ||
         !isInSkyblock() ||
-        getWorldName() === KUUDRA ||
+        !isInFishingWorld(getWorldName()) ||
         !hasFishingRodInHotbar() ||
         allOverlaysGui.isOpen()
     ) {